Fix interrupt/exceptions/irq/isr code

+++ include/serial.h
@@ -4,6 +4,6 @@
 
 uint8_t inb(unsigned int port);
 void outb(unsigned int port,unsigned char value);
-void set_int(int i, uint32_t addr);
+void set_int(int i, uint64_t addr);
 void send_eoi(uint8_t irq);
 #endif
